Move proxied methods to InputProxy
